requirements.txt
streamlit
crewai
openai
anthropic
google-generativeai
